# Performance optimisation for time-series `TableMixin`