--- a/CHANGES.md
+++ b/CHANGES.md
@@ -16,7 +16,7 @@
 - Added missing `proxyCatalogItemUrl` to GeoJson, Shapefile, Gltf and AssImp catalog items.
 - Added support for `OpenDataSoftCatalogGroup` with more than 100 datasets.
 - Added `refreshIntervalTemplate` to `OpenDataSoftCatalogItemTraits` - this can be used to set `refreshInterval` using Mustache template rendered on ODS Dataset JSON object.
-
+- Performance optimisation for time-series `TableMixin`
 - [The next improvement]
 
 #### 8.2.17 - 2022-09-23