### Performance optimisation for time-series `TableMixin`

25-70% reduction in time to process time-series datasets. The reduction will depend on

- Number of features
- Number of samples per feature
- Number of values per sample (columns) 
- Complexity of styling

### How

There are two changes
- Got rid of an `Array.reduce` and `Array.some` in `TableMixin.discreteTimes`  - they are both super slow and unnecessary 
- Added this new object called `PreSampledProperty` in `createLongitudeLatitudeFeaturePerId` which creates `SampledProperty` after all samples have been collected (using `addSamples`) - instead of creating a `SampledProperty` before and calling `addSample` for each individual sample

### Evaluation

Notes:

- "Complex Styling"
  - "No" = static color points
  - "Yes" = scalar color, label and trail styling
- Only one sample for before and after

Features | Samples per Feature | Values (columns) | Complex Styling | Before - create features (ms) | Before - process dates (ms) | After - create features (ms) | After - process dates (ms) | Reduction
-- | -- | -- | -- | -- | -- | -- | -- | --
25 | 1000 | 0 | No | 481.86 | 753.53 | 485.96 | 189.58 | 45%
25 | 1000 | 1 | No | 1130.53 |   | 587.07 |   | 48%
25 | 1000 | 5 | No | 1234.29 | 807.72 | 695.99 | 263.89 | 53%
25 | 5000 | 0 | No | 5467.94 | 3580.98 | 5195.65 | 650.63 | 35%
25 | 5000 | 1 | No | 11104.02 |   | 5158.35 |   | 54%
25 | 5000 | 5 | No | 11823.63 | 3791.81 | 6085.02 | 863.65 | 56%
25 | 10000 | 0 | No | 15941.14 | 6921.36 | 16176.54 | 1257.6 | 24%
25 | 10000 | 1 | No | 22648.96 |   | 12938.6 |   | 43%
25 | 10000 | 5 | No | 25482.37 | 7683.67 | 14456.68 | 1685.54 | 51%
25 | 10000 | 5 | Yes | 41670.59 | 7683.67 | 20120.97 | 1685.54 | 56%
50 | 1000 | 0 | No | 930.23 | 1419.97 | 949.1 | 307.61 | 47%
50 | 1000 | 1 | No | 2148.63 |   | 1141.09 |   | 47%
50 | 1000 | 5 | No | 2458.27 | 1613.52 | 1290.75 | 370.77 | 59%
50 | 1000 | 5 | Yes | 4778.2 | 1613.52 | 1746.38 | 370.77 | 67%
50 | 5000 | 0 | No | 10724.31 | 6960.78 | 12101.56 | 1384.63 | 24%
50 | 5000 | 1 | No | 23199.54 |   | 10273.39 |   | 56%
50 | 5000 | 5 | No | 24477.09 | 7991.16 | 10987.49 | 1580.27 | 61%
50 | 5000 | 5 | Yes | 44043.07 | 7991.16 | 17683.62 | 1580.27 | 63%
50 | 10000 | 0 | No | 31321.46 | 13674.49 | 30388.4 | 2204.95 | 28%
50 | 10000 | 1 | No | 44439.66 |   | 26685.38 |   | 40%
50 | 10000 | 5 | No | 54030.81 | 16477.63 | 28769.36 | 3427.89 | 54%
50 | 10000 | 5 | Yes | 84553.18 | 16477.63 | 35449.76 | 3427.89 | 62%
100 | 1000 | 0 | No | 1731.99 | 2975.85 | 1749.92 | 527.7 | 52%
100 | 1000 | 1 | No | 4730.68 |   | 2235.43 |   | 53%
100 | 1000 | 5 | No | 4906.03 | 3008.17 | 2586.72 | 648.88 | 59%
100 | 5000 | 0 | No | 21303.13 | 13760.61 | 21003.67 | 2266.48 | 34%
100 | 5000 | 1 | No | 42416.54 |   | 20516.9 |   | 52%
100 | 5000 | 5 | No | 48701.25 | 14616.51 | 22852.89 | 3236.63 | 59%
100 | 10000 | 0 | No | 62577.95 | 27732.8 | 59715.75 | 4398.23 | 29%
100 | 10000 | 1 | No | 87747.35 |   | 53672.81 |   | 39%
100 | 10000 | 5 | No | 100822.79 | 29820.78 | 57684.57 | 6650.73 | 51%


### Test me

How should reviewers test this? (Hint: If you want to provide a test catalog item, [create a Gist](https://gist.github.com/) of its catalog JSON, add its url and your branch name to this url: `http://ci.terria.io/<branch name>/#clean&<raw url of your gist>` , and then paste that in the body of this PR)

### Checklist

- [x] There are unit tests to verify my changes are correct or unit tests aren't applicable (if so, write quick reason why unit tests don't exist)
- [ ] I've updated relevant documentation in `doc/`.
- [x] I've updated CHANGES.md with what I changed.
- [ ] I've provided instructions in the PR description on how to test this PR.